Struct blobwar::configuration::Configuration

pub struct Configuration<'a> {
    pub current_player: bool,
    /* private fields */
}

Game state. We know, who should play, what is the board and where every blob is located.

Fields
current_player: bool

Who should play now.

Implementations

impl<'a> Configuration<'a>

pub fn new(board: &'a Board) -> Self

Create an initial game configuration out of given Board.

pub fn apply_movement(&mut self, movement: &Movement)

Play given move on self.

pub fn play(&self, movement: &Movement) -> Self

Create a new Configuration by playing given Movement on self.

pub fn skip_play(&self) -> Self

Create a new Configuration by skipping turn on self.

pub fn free_position_at(&self, position: u8) -> bool

Does target player has blob at target position set ?

pub fn empty_cells(&self) -> impl Iterator<Item = u8>

Iterate on all Position(s) of empty cells.

pub fn value(&self) -> i8

Return the configuration value (#other_player - #current_player)

pub fn check_move(&self, movement: &Movement) -> bool

Return if given movement is correct for current configuration.

pub fn battle<T: Strategy, U: Strategy>(&mut self, player_one: T, player_two: U)

Play a match between the given players starting from current Configuration.

pub fn movements<'b>(&'b self) -> impl 'b + Iterator<Item = Movement>

Iterate on all possible moves.

pub fn serialize(&self) -> String

Serialize Configuration into a String. Use in communications with sub-processes.

pub fn deserialize(string: &str, board: &'a Board) -> Self

Deserialize given String into a Configuration. You need to deserialize the Board first.