Module blobwar::strategy

We provide here structs for all possible kinds of players and AI.

Re-exports

pub use self::human::Human;

pub use self::network::NetworkPlayer;

pub use self::greedy::Greedy;

pub use self::minmax::min_max_anytime;

pub use self::minmax::MinMax;

pub use self::alphabeta::alpha_beta_anytime;

pub use self::alphabeta::AlphaBeta;

pub use self::iterative::IterativeDeepening;

pub use self::iterative::IterativeStrategy;

Modules

alphabeta
Alpha - Beta algorithm.

greedy
Dumb greedy algorithm.

human
Human player.

iterative
Generic iterative deepening strategies (with variable algorithms).

minmax
Implementation of the min max algorithm.

network
Network player (server side)

Traits

Strategy
To be a strategy you need to be able to compute the next move.