Trait blobwar::strategy::Strategy

pub trait Strategy: Display {
    fn compute_next_move(
        &mut self, 
        configuration: &Configuration<'_>
    ) -> Option<Movement>;
}

To be a strategy you need to be able to compute the next move.

Required methods

fn compute_next_move(
&mut self,
configuration: &Configuration<'_>
) -> Option<Movement>

Take current Configuration and return what to do next. None if no move is possible.